lessons_screen.dart finished (showing the lessons list sorted)

+++ marcelo_esser_test/lib/dao/lessons_dao.dart
@@ -111,7 +111,7 @@ class LessonsDao {
 
       summaries.add(
         Summary(
-          percentage: percentage.toInt(),
+          percentage: percentage,
           id: value[_summaryId],
         ),
       );
